+++ R/fct_export.R
@@ -1,9 +1,0 @@
-#' export_pdf
-#'
-#' @description export PDFs from a collection
-#'
-#' @return The return value, if any, from executing the function.
-#'
-#' @noRd
-export_pdf <- function() {
-}