+++ tests/testthat/test-mod_collections.R
@@ -1,8 +1,8 @@
 testServer(
   mod_collections_server,
-  # Add here your module params
-  args = list()
-  , {
+  args = list(
+    credentials = reactive(data.frame(UserID = NA_real_, APIKey = ""))
+  ), {
     ns <- session$ns
     expect_true(
       inherits(ns, "function")
@@ -25,7 +25,7 @@ testServer(
     # - Testing output
     # expect_true(inherits(output$tbl$html, "html"))
 })
- 
+
 test_that("module ui works", {
   ui <- mod_collections_ui(id = "test")
   golem::expect_shinytaglist(ui)
@@ -35,4 +35,4 @@ test_that("module ui works", {
     expect_true(i %in% names(fmls))
   }
 })
- 
+